+++ typings/RequiresUpdateMeta.ts
@@ -1,0 +1,7 @@
+
+export default interface RequiresUpdateMeta {
+  fileUrl: string,
+  hasFile: boolean;
+  needsUpdate: boolean;
+  lastModified: Date | null,
+}